Refactor Edit module
Move Provider and Bandwidth adapter methods to a PHP trait to reuse the
behavior in SetPppoeParameters.
